GGA closed shell seems to work fine now.
The code is not really changed, however the H2 polarizability test
converges to the same result for both the older and the newer version.
The only noticeable modification is in XCPotential.cpp, where I update
rho_gs and rho_pert to one another. I had to remove this "symmetric"
update for open shell to guarantee that both alpha and beta densities
are defined on the same grid. If this is not yhe reason why the test
now works, I can only guess that some improvements in mrcpp has
affected the gradient/potential calculations.